Fix mockall_derive's unit tests
